Change use of comports() list for POSIX platforms
When run on Linux, this script would barf endlessly because the "name" of a comport is not the same as the "device" of the comport list object. When you go to try and open a device that isn't listed as "/dev/ttyUSB0", it would crash out rather vaguely with "Cannot connect to device".